Change to cast a and b to u128 so we test u128 = u64 + u64
Also save benchmark.fast.asm which was generated with:
  $ objdump --source -d -M intel ./zig-cache/test > benchmark.fast.asm

Update README.md to include how I created the asm and what the actaul
asm code of the loop.

+++ benchmark.zig
@@ -181,7 +181,7 @@ test "benchmark.add" {
             var pA: *volatile u64 = &pSelf.a;
             var pB: *volatile u64 = &pSelf.b;
             var pR: *volatile u128 = &pSelf.r;
-            pR.* = pA.* + pB.*;
+            pR.* = u128(pA.*) + u128(pB.*);
             //sfence();
             //@fence(AtomicOrder.Release); // Generates no type of fence, expected sfence
             //@fence(AtomicOrder.AcqRel); // Generates no type of fence, expected ??
